requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic-settings==2.4.0
easyocr==1.7.2
pillow==10.4.0
numpy==1.26.4
python-multipart==0.0.9
diffusers>=0.30.0
transformers>=4.44.0
accelerate>=0.33.0
imageio[ffmpeg]>=2.34.0
